+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Do not attach headers if Span is NoOp ([#1112](https://github.com/getsentry/sentry-dart/pull/1112))
+
 ## 6.16.0
 
 ### Features